@@ -1,28 +1,29 @@
 source 'https://rubygems.org'
 
-gem 'rails', '4.2.0'
+ruby '2.7.1'
+gem 'rails', '~> 6.0.3', '>= 6.0.3.4'
 gem 'pg'
-gem 'sass-rails', '~> 5.0'
-gem 'uglifier', '>= 1.3.0'
-gem 'coffee-rails', '~> 4.1.0'
+gem 'puma', '~> 4.1'
+gem 'sass-rails', '>= 6'
+gem 'webpacker', '~> 4.0'
+gem 'uglifier'#, '>= 1.3.0'
+gem 'coffee-rails'#, '~> 4.1.0'
 gem 'haml-rails'
 gem 'jquery-rails'
-gem 'jquery-ui-rails'
+gem 'jquery-ui-rails', '5.0.3'
 gem 'turbolinks'
-gem 'jbuilder', '~> 2.0'
-gem 'delayed_job', '~> 4.0.6'
-gem 'delayed_job_active_record', '~> 4.0.3'
-gem "highcharts-rails", "~> 3.0.0"
+gem 'jbuilder'#, '~> 2.0'
+gem 'delayed_job'#, '~> 4.0.6'
+gem 'delayed_job_active_record'#, '~> 4.0.3'
+gem "highcharts-rails"#, "~> 3.0.0"
 gem 'd3_rails'
 gem 'c3-rails'
 gem 'faye-authentication', github: 'branch14/faye-authentication'
 gem 'devise'
-# sprockets is locked so that it does not update by accident to v3 which will
-# crash capistrano deployment currently
-gem 'sprockets', '2.12.3'
-
-gem 'sdoc', '~> 0.4.0', group: :doc
-
+gem 'sprockets'#, '~> 3.0'
+gem 'mime-types'
+gem 'sdoc', '1.1.0', group: :doc
+gem 'sqlite3'#, '~> 1.3.6'
 # optional
 gem 'active_skin'
 
@@ -41,32 +42,42 @@ gem 'fog-aws', require: 'fog/aws'
 gem 'unicorn'
 gem 'activeadmin', github: 'activeadmin/activeadmin'
 gem 'rails_config'
-gem 'activeadmin-dragonfly', github: 'stefanoverna/activeadmin-dragonfly'
+gem 'activeadmin-dragonfly'#, github: 'stefanoverna/activeadmin-dragonfly'
 gem 'acts-as-taggable-on'
 
 group :production do
-  gem 'airbrake', '~> 5.0'
+  gem 'airbrake'#, '~> 5.0'
 end
 
 group :development do
   gem 'annotator'
-  gem 'capistrano',         '~> 3.1.0'
-  gem 'capistrano-rbenv',   '~> 2.0.1'
-  gem 'capistrano-bundler', '~> 1.1.1'
-  gem 'capistrano-rails',   '~> 1.1.1'
+  gem 'capistrano',         '~> 3.4.0'
+  gem 'capistrano-rbenv',   '~> 2.1.0'
+  gem 'capistrano-bundler', '~> 1.6.0'
+  gem 'capistrano-rails',   '~> 1.5.0'
 end
 
 group :development, :test do
+  gem 'listen', '~> 3.2'
   gem 'spring'
-  gem 'web-console', '~> 2.0'
+  gem 'web-console'#, '~> 2.0'
   gem 'pry-rails'
   gem 'byebug'
   #gem 'disable_assets_logger'
 
   gem 'database_cleaner'
   gem 'rspec-rails'
-  gem 'rspec_junit_formatter', '0.2.2'
+  gem 'rspec_junit_formatter'#, '0.2.2'
   gem 'spring-commands-rspec' # TODO check if needed
   gem 'rack-test'
   gem 'factory_girl_rails'
 end
+
+gem 'config'
+gem 'zeitwerk'
+gem 'rake'
+gem 'dragonfly'
+gem "bourbon"
+gem 'figaro'
+gem 'activestorage-backblaze'
+gem "aws-sdk-s3", require: false
