@@ -1,2 +1,2 @@
-class Setting < ActiveRecord::Base
+class Setting < ApplicationRecord
 end
