@@ -1,7 +1,7 @@
 # the require is needed since we're monkey patching delayed job
 # with ap/model/delayed_job, but since it undermines rails' autoload
 # we'll have to restart the server after editing ap/model/delayed_job
-require Rails.root.join 'app', 'models', 'delayed_job'
+# require Rails.root.join 'app', 'models', 'delayed_job'
 
 ActiveAdmin.register Delayed::Job, as: "Job" do
 
@@ -54,9 +54,9 @@ ActiveAdmin.register Delayed::Job, as: "Job" do
     column :run_at
     column :attempts
     column :queue
-    column :handler do |job|
-      job.display_handler
-    end
+    # column :handler do |job|
+    #   job.display_handler
+    # end
     column :locked_at
     column :locked_by
     column :last_error do |job|
@@ -67,9 +67,9 @@ ActiveAdmin.register Delayed::Job, as: "Job" do
   end
 
   scope :all
-  scope :failed
-  scope :audio
-  scope :trigger
-  scope :mail
+  # scope :failed
+  # scope :audio
+  # scope :trigger
+  # scope :mail
 
 end
