@@ -127,7 +127,7 @@ ActiveAdmin.register Venue do
       f.input :long
       f.input :options, input_html: { rows: 6 },
               hint: "Boolean flags will be set/overriden by the checkboxes below."
-      f.input :facebook_live_url, hint: "Facebook terminates streams longer than 4 hours. E.g. rtmp://live-api-a.facebook.com:80/rtmp/1596393490422675?ds=1&s_l=1&a=ATjm8SZd5qCK1K8W"
+      # f.input :facebook_live_url, hint: "Facebook terminates streams longer than 4 hours. E.g. rtmp://live-api-a.facebook.com:80/rtmp/1596393490422675?ds=1&s_l=1&a=ATjm8SZd5qCK1K8W"
       # FIXME selected options are not properly checkmarked
       f.input :flags, as: :check_boxes, collection: Venue::FLAGS,
               member_label: :humanize,
