@@ -35,7 +35,7 @@ Rails.application.configure do
   # Checks for improperly declared sprockets dependencies.
   # Raises helpful error messages.
   config.assets.raise_runtime_errors = true
-
+  config.active_storage.service = :backblaze
   # Raises error for missing translations
   # config.action_view.raise_on_missing_translations = true
 
